hw6

+++ src/main/java/org/example/Client.java
@@ -1,23 +1,23 @@
 package org.example;
 
 public class Client {
-    private int id;
+    private long id;
     private String name;
 
     public Client() {
     }
 
-    public Client(int id, String name) {
+    public Client(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
